Add files via upload
As part of my 2nd year engineering labs at Oxford, I programmed a simple android app along with two team members. The app has two views: 
----------------
View 1 allows the user to take a photo, which is automatically saved along with the geolocation. To create View 1, we had to do the following:
-Define application permissions: 
--Storage access
--Vibrations
--Internet
--Location Services
-Initialize Google Maps
-Set up Shared Preferences
-Create a Location Manager & Location Listener
-Set up Take Picture button:
 --Create file paths for Images
 --Define intents to take photos 
 --Store geolocations

--------------
View 2 allows the user to view the entire journey as a playback - each location is plotted as a pin on the map fragment, and as each pin appears, so does the photo taken at that location. The app also draws a line between each location to show the entire journey. To create view 2, we had to:
-Retrieve info from View 1 (SharedPreferences)
--Retrieve image fi